requirements.txt
streamlit>=1.20.0
face_recognition==1.3.0
opencv-python
numpy
Pillow
